+++ CHANGELOG.md
@@ -17,6 +17,7 @@ in 8.x, 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 ### 2026-02-xx
   * Updated PackingSolver dependency
   * Improved Smart Stretch Tool performance
+  * Improved management of locked states in the Smart Stretch tool
   * Improved Outliner Deep Rename Parts feature
 ### 2026-02-06
   * Updated PackingSolver dependency